move first fetch actions to the App component
