+++ src/App.jsx
@@ -1,5 +1,6 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import { useDispatch } from 'react-redux';
 import { Footer } from './app/Footer';
 import { MovieRoute } from './features/movies/MovieRoute';
 
@@ -8,17 +9,33 @@ import { Home } from './app/Home';
 import { Login } from './app/Login';
 import { SignUp } from './app/SignUp';
 import { ActionRoute } from './features/movies/ActionRoute';
+import {
+  fetchActionsAction,
+  fetchAnimationsAction,
+  fetchComediesAction,
+  fetchPopularsAction,
+} from './app/store/actions/moviesAction';
 
-export const App = () => (
-  <Router>
-    <Navbar />
-    <Switch>
-      <Route path='/movie/:id' component={MovieRoute} />
-      <Route path='/login' component={Login} />
-      <Route path='/signup' component={SignUp} />
-      <Route path='/action' component={ActionRoute} />
-      <Route exact path='/' component={Home} />
-    </Switch>
-    <Footer />
-  </Router>
-);
+export const App = () => {
+  const dispatch = useDispatch();
+  useEffect(() => {
+    dispatch(fetchComediesAction());
+    dispatch(fetchAnimationsAction());
+    dispatch(fetchPopularsAction());
+    dispatch(fetchActionsAction());
+  }, []);
+
+  return (
+    <Router>
+      <Navbar />
+      <Switch>
+        <Route path='/movie/:id' component={MovieRoute} />
+        <Route path='/login' component={Login} />
+        <Route path='/signup' component={SignUp} />
+        <Route path='/action' component={ActionRoute} />
+        <Route exact path='/' component={Home} />
+      </Switch>
+      <Footer />
+    </Router>
+  );
+};
